fix: デプロイを完了して人力評価ができるようにする
