+++ rag_chain_config.yaml
@@ -1,4 +1,3 @@
 llm_endpoint_name: databricks-dbrx-instruct
-source_table_name: dev.rach_db.rach_documentation
 vector_search_endpoint_name: vs_endpoint
 vector_search_index_name: dev.rach_db.rach_documentation_vs_index
